Create LEDSweepingPattern.s
Program that turns on two LEDR lights at a time on the DE1- SoC board. First, the lights LEDR9 and LEDR0  turn on, then LEDR8 and LEDR1, then LEDR7 and LEDR2, and so on. When it reaches LEDR5 and LEDR4, the direction reverses. Only two LEDR lights are ever on at one time. The effect is two lights sweeping first towards the centre, then outwards towards the edges, and so on.